requirements.txt
beautifulsoup4>=4.12.0
google-genai>=1.0.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0

# Optional: for MathML→LaTeX in Transcription (convert_mathml_to_latex.py). If missing, falls back to Node mathml-to-latex.
# Install Ruby gem: gem install plurimath
